It's going a lot faster now.  Added wareRefs and did some redesign

+++ src/main/java/com/assaydepot/result/WareRef.java
@@ -4,17 +4,17 @@ import java.util.List;
 
 public class WareRef extends BaseResult {
 
-	private Float price;
+	private Double price;
 	private Integer turnAroundTime;
 	private String Type;
 	private String snippet;
 	private List<String> providerIds;
 	private List<String> providerNames;
-	private Float score;
-	public Float getPrice() {
+	private Double score;
+	public Double getPrice() {
 		return price;
 	}
-	public void setPrice(Float price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 	public Integer getTurnAroundTime() {
@@ -47,10 +47,10 @@ public class WareRef extends BaseResult {
 	public void setProviderNames(List<String> providerNames) {
 		this.providerNames = providerNames;
 	}
-	public Float getScore() {
+	public Double getScore() {
 		return score;
 	}
-	public void setScore(Float score) {
+	public void setScore(Double score) {
 		this.score = score;
 	}
